requirement.txt
alabaster==0.7.12
alembic==0.8.10
altair==1.2.1
apache-airflow==1.10.2
asn1crypto==0.22.0
atari-py==0.1.7
attrs==19.1.0
audioread==2.1.4
av==0.3.3
aws-xray-sdk==0.95
awscli==1.16.17
Babel==2.6.0
backcall==0.1.0
backports.functools-lru-cache==1.4
backports.weakref==1.0rc1
bcrypt==3.1.6
beautifulsoup4==4.6.0
bleach==2.1.4
blinker==1.4
bokeh==0.12.13
boto==2.48.0
boto3==1.7.84
botocore==1.10.84
Box2D==2.3.2
Box2D-kengz==2.3.3
bresenham==0.2
bz2file==0.98
cachetools==3.1.0
cassandra-driver==3.11.0
certifi==2019.3.9
cffi==1.11.2
cgroupspy==0.1.6
chardet==3.0.4
click==6.7
cloudpickle==0.2.2
colorama==0.3.9
colour==0.1.5
conda==4.6.9
configparser==3.5.3
cookies==2.2.1
croniter==0.3.29
cryptography==2.1.4
cycler==0.10.0
Cython==0.29.6
dask==0.16.1
decorator==4.0.11
defusedxml==0.5.0
dill==0.2.7.1
docker==3.7.2
docker-pycreds==0.4.0
docutils==0.14
dronekit==2.9.1
ecdsa==0.13
entrypoints==0.2.3
enum34==1.1.6
et-xmlfile==1.0.1
eventlet==0.22.0
fastcache==1.0.2
Flask==0.12.4
Flask-Admin==1.5.2
Flask-AppBuilder==1.12.1
Flask-Babel==0.12.2
Flask-Bcrypt==0.7.1
Flask-Caching==1.3.3
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SocketIO==2.9.2
Flask-SQLAlchemy==2.3.2
flask-swagger==0.2.13
Flask-WTF==0.14.2
freezegun==0.3.11
funcsigs==1.0.0
future==0.16.0
gensim==3.4.0
gevent==1.4.0
gitdb2==2.0.5
GitPython==2.1.11
glfw==1.7.1
google-auth==1.6.3
greenlet==0.4.12
gunicorn==19.9.0
gym==0.9.6
h5py==2.7.1
heapdict==1.0.0
hmsclient==0.1.1
html2text==2018.1.9
html5lib==1.0.1
idna==2.6
imageio==2.1.2
imagesize==1.1.0
inflection==0.3.1
ipykernel==4.9.0
ipython==6.5.0
ipython-genutils==0.2.0
ipython-sql==0.3.9
ipywidgets==7.0.5
iso8601==0.1.12
itsdangerous==0.24
jdc==0.0.8
jdcal==1.3
jedi==0.10.2
jieba==0.39
Jinja2==2.10
jira==2.0.0
jmespath==0.9.3
joblib==0.11
json-merge-patch==0.2
jsondiff==1.1.1
jsonpickle==1.1
jsonschema==2.6.0
jupyter-client==5.2.4
jupyter-core==4.4.0
jupyterlab==0.35.4
jupyterlab-server==0.2.0
Keras==2.0.9
kubernetes==9.0.0
librosa==0.5.1
llvmlite==0.20.0
locket==0.2.0
lockfile==0.12.2
lxml==4.1.1
Mako==1.0.7
Markdown==2.6.9
MarkupSafe==1.0
matplotlib==2.1.0
mistune==0.8.3
mock==2.0.0
mongomock==3.16.0
monotonic==1.2
moto==1.3.5
moviepy==0.2.3.2
msgpack==0.5.6
msgpack-python==0.4.8
mujoco-py==1.50.1.59
mysqlclient==1.4.2.post1
nbconvert==5.4.0
nbformat==4.4.0
networkx==1.11
nltk==3.2.5
nose==1.3.7
nose-ignore-docstring==0.2
nose-timer==0.7.5
notebook==5.7.0
ntlm-auth==1.2.0
numba==0.35.0
numexpr==2.6.4
numpy==1.12.1
oauthlib==2.0.6
olefile==0.44
openpyxl==2.5.0b1
ordereddict==1.1
packaging==16.8
pandas==0.23.3
pandocfilters==1.4.1
parameterized==0.7.0
paramiko==2.4.2
parso==0.3.4
partd==0.3.8
patsy==0.4.1
pbr==3.1.1
pendulum==1.4.4
pexpect==4.3.1
pickleshare==0.7.4
Pillow==4.2.1
pinyin==0.4.0
plotly==2.0.15
pomegranate==0.9.0
prettytable==0.7.2
prometheus-client==0.3.1
prompt-toolkit==1.0.15
protobuf==3.5.1
psutil==5.4.0
psycopg2==2.7.4
ptyprocess==0.5.2
py4j==0.10.7
pyaml==18.11.0
pyasn1==0.4.4
pyasn1-modules==0.2.4
pycosat==0.6.3
pycparser==2.18
pycryptodome==3.8.0
pycurl==7.43.0
pydot==1.2.4
pydotplus==2.0.2
pyglet==1.3.0
Pygments==2.2.0
pygpu==0.7.5
PyHive==0.6.1
PyJWT==1.5.3
pymavlink==2.2.8
pymc3==3.2
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==17.4.0
pyparsing==2.2.0
pyrsistent==0.14.9
pysftp==0.2.9
PySocks==1.6.8
pyspark==2.4.0
python-crfsuite==0.9.2
python-daemon==2.1.2
python-dateutil==2.6.1
python-editor==1.0.4
python-engineio==2.0.2
python-jose==2.0.2
python-nvd3==0.15.0
python-slugify==3.0.2
python-socketio==1.8.4
python-speech-features==0.6
python3-openid==3.1.0
pytz==2017.3
pytzdata==2018.9
PyWavelets==0.5.2
pywinrm==0.3.0
PyYAML==3.12
pyzmq==17.1.2
qds-sdk==1.11.0
rednose==1.3.0
requests==2.21.0
requests-mock==1.5.2
requests-ntlm==1.1.0
requests-oauthlib==0.8.0
requests-toolbelt==0.9.1
resampy==0.2.0
responses==0.10.6
rsa==3.4.2
ruamel-yaml==0.11.14
s3transfer==0.1.13
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==0.19.1
seaborn==0.8.1
SeekWell==0.1
Send2Trash==1.5.0
sentinels==1.0.0
setproctitle==1.1.10
Shapely==1.6.4.post1
simplegeneric==0.8.1
SimpleWebSocketServer==0.1.1
six==1.11.0
smart-open==1.5.6
smmap2==2.0.5
snakebite==2.11.0
snowballstemmer==1.2.1
snownlp==0.12.3
SoundFile==0.9.0.post1
Sphinx==2.0.0
sphinx-argparse==0.2.5
Sphinx-PyPI-upload==0.2.1
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.1
sphinxcontrib-httpdomain==1.7.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.1
SQLAlchemy==1.2.18
sqlparse==0.3.0
statsmodels==0.8.0
sympy==1.0
tabulate==0.8.2
tenacity==4.8.0
tensorflow==1.3.0
terminado==0.8.1
termstyle==0.1.11
testpath==0.3.1
text-unidecode==1.2
Theano==1.0.1
thrift==0.11.0
toolz==0.8.2
torch==0.4.0
torchvision==0.2.1
tornado==4.5.3
tqdm==4.11.2
traitlets==4.3.2
tweepy==3.5.0
twython==3.6.0
tzlocal==1.5.1
udacity-pa==0.2.7
unicodecsv==0.14.1
urllib3==1.22
utm==0.4.2
uvloop==0.8.1
vega==0.4.4
vincent==0.4.4
wcwidth==0.1.7
webencodings==0.5
websocket-client==0.56.0
websockets==4.0.1
Werkzeug==0.14.1
widgetsnbextension==3.1.0
wordcloud==1.4.1
wptools==0.4.7
wrapt==1.11.1
WTForms==2.2.1
xlrd==1.1.0
xmltodict==0.12.0
zope.deprecation==4.4.0

